+++ src/pages/CopyMcDataProgress.svelte
@@ -1,6 +1,6 @@
 <script>
     import TransitionWrapper from "./TransitionWrapper.svelte";
-    import CopyMcDataProgressScreen from "../components/main/CopyMcDataProgressScreen.svelte";
+    import CopyMcDataProgressScreen from "../components/clone-mc-data/CopyMcDataProgressScreen.svelte";
 </script>
 
 <TransitionWrapper>
