+++ src/pages/FirstInstall.svelte
@@ -1,6 +1,6 @@
 <script>
   import TransitionWrapper from "./TransitionWrapper.svelte";
-  import FirstInstallScreen from "../components/main/FirstInstallScreen.svelte";
+  import FirstInstallScreen from "../components/clone-mc-data/FirstInstallScreen.svelte";
 </script>
 
 <TransitionWrapper>
